Added tracking calculations helper
